+++ MiniTwitch.Helix/Internal/Json/Converters.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text.Json.Serialization;
 using System.Text.Json;
 using System.Text;
+using MiniTwitch.Helix.Responses;
 
 namespace MiniTwitch.Helix.Internal.Json;
 
@@ -111,4 +112,30 @@ internal class TimeSpanToSeconds : JsonConverter<TimeSpan>
         => throw new NotSupportedException("Converter should only be used for writing");
 
     public override void Write(Utf8JsonWriter writer, TimeSpan value, JsonSerializerOptions options) => writer.WriteNumberValue((int)value.TotalSeconds);
+}
+
+/// <summary>
+/// Converts <see cref="ConduitTransport"/>
+/// </summary>
+public class ConduitTransportConverter : JsonConverter<ConduitTransport>
+{
+    public override void Write(Utf8JsonWriter writer, ConduitTransport value, JsonSerializerOptions options)
+        => throw new NotImplementedException("This converter is read-only");
+
+    public override ConduitTransport? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    {
+        var root = JsonDocument.ParseValue(ref reader).RootElement;
+
+        if (root.TryGetProperty("method", out var method))
+        {
+            return method.GetString() switch
+            {
+                "webhook" => root.Deserialize<ConduitTransport.Webhook>(),
+                "websocket" => root.Deserialize<ConduitTransport.WebSocket>(),
+                _ => throw new JsonException($"Unknown method: {method.GetString()}")
+            };
+        }
+
+        return null;
+    }
 }
